backend/requirements.txt
asgiref==3.9.1
attrs==25.4.0
black==25.12.0
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
defusedxml==0.7.1
distlib==0.4.0
Django==5.2.4
django-bootstrap5==25.3
django-filter==25.2
django-sequences==3.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-spectacular==0.29.0
drf-spectacular-sidecar==2026.1.1
filelock==3.20.2
identify==2.6.15
idna==3.11
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mypy_extensions==1.1.0
nodeenv==1.10.0
oauthlib==3.3.1
packaging==25.0
pathspec==1.0.2
platformdirs==4.5.1
pluggy==1.6.0
pre_commit==4.5.1
psycopg2-binary==2.9.11
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pyTelegramBotAPI==4.29.1
pytest==8.4.1
pytest-django==4.11.1
python-dotenv==1.2.1
python3-openid==3.2.0
pytokens==0.3.0
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
ruff==0.14.10
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.36.0
